separated all types from URLSession.swift into interceptablesession etc., removed test example file, fixed all but two annoying warnings
